Refactored seeding for better readability and scalability
