+++ src/seed/pool.ts
@@ -1,0 +1,60 @@
+export const firstNames: string[] = [
+  "Tokino",
+  "Kanata",
+  "Watame",
+  "Roboco",
+  "Suisei",
+  "Azki",
+  "Aki",
+  "Akai",
+  "Matsuri",
+  "Mei",
+];
+export const lastNames: string[] = [
+  "Amane",
+  "Kiryuu",
+  "Sora",
+  "Haato",
+  "Rosenthal",
+  "Sung086",
+];
+export const addresses: string[] = [
+  "Home",
+  "Bangkok somewhere",
+  "USA",
+  "OHIO",
+  "Chiangmai",
+  "Isekai",
+  "OtakuRoom",
+];
+export const petTypes: string[] = [
+  "Dog",
+  "Cat",
+  "Goldfish",
+  "Panda",
+  "Snake",
+  "Bat",
+  "Lizard",
+  "Hamster",
+];
+
+export const imageUris: string[] = [
+  "https://i.ppy.sh/547c489981845883d37a525d50d892d8c3d490e0/68747470733a2f2f692e696d6775722e636f6d2f5430586455414e2e6a7067",
+  "https://p4.wallpaperbetter.com/wallpaper/771/366/387/anime-anime-girls-hololive-kiryu-coco-redhead-hd-wallpaper-preview.jpg",
+  "https://img-9gag-fun.9cache.com/photo/aZrDB5Q_460s.jpg",
+  "https://preview.redd.it/8p4t0a95ax161.jpg?auto=webp&s=9c9db601dfdf217647380cbffe44d7eaf0efa47b",
+  "https://gamerbraves.sgp1.cdn.digitaloceanspaces.com/2022/02/Rushia_FI2.jpg",
+  "https://hololive.hololivepro.com/wp-content/uploads/2022/06/%E5%8D%9A%E8%A1%A3%E3%81%93%E3%82%88%E3%82%8A_WAO_jk-1.png",
+  "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQGYXihhLktMheIVxxgkU3ouhnXf9F9wbxlhQ&usqp=CAU",
+  "https://cdn.donmai.us/original/98/83/98834cd37301a94a6a524d94f1ff5a88.jpg",
+  "https://static.wikia.nocookie.net/omniversal-battlefield/images/b/bd/Council.jpg/revision/latest?cb=20220420023638",
+  "https://picsum.photos/200",
+];
+
+export const reviewTexts: string[] = [
+  "My dog said it was great.",
+  "good... I guess?",
+  "dog shit",
+  "Pretty nice",
+  "Skibidi pop pop pop pop ye ye ye ye",
+];
