+++ src/seed/util.ts
@@ -1,0 +1,10 @@
+export function getMultipleRandom(arr: string[], num: number) {
+  const shuffled = [...arr].sort(() => 0.5 - Math.random());
+
+  return shuffled.slice(0, num);
+}
+
+export function getRandomIntFromInterval(min: number, max: number) {
+  // min and max included
+  return Math.floor(Math.random() * (max - min + 1) + min);
+}
